Complete directory change for the webserver Switched to python library structure All import are in __init__.py of the app library
